+++ packages/zkp-gen-iden3/src/enums/operators.ts
@@ -1,0 +1,9 @@
+export enum ZKP_OPERATORS {
+  SKIP = '0',
+  EQUALS = '1',
+  LESS_THAN = '2',
+  GREATER_THAN = '3',
+  IN = '4',
+  NOTIN = '5',
+  NOT_EQUALS = '6',
+}