+++ packages/zkp-gen-iden3/src/helpers/index.ts
@@ -1,0 +1,3 @@
+export * from './binary-helpers'
+export * from './file-helpers'
+export * from './state-v2-helpers'